ui: Move timeline-internal traits into separate module
