+++ crates/matrix-sdk-ui/src/timeline/read_receipts.rs
@@ -24,10 +24,8 @@ use ruma::{
 use tracing::{error, warn};
 
 use super::{
-    compare_events_positions,
-    event_item::EventTimelineItemKind,
-    inner::{RoomDataProvider, TimelineInnerState},
-    rfind_event_by_id, EventTimelineItem, RelativePosition, TimelineItem,
+    compare_events_positions, event_item::EventTimelineItemKind, inner::TimelineInnerState,
+    rfind_event_by_id, traits::RoomDataProvider, EventTimelineItem, RelativePosition, TimelineItem,
 };
 
 struct FullReceipt<'a> {
